requirements.txt
telethon
aiogram==3.10.0
flask
requests
python-dotenv
